+++ src/block-kit/Section.tsx
@@ -1,10 +1,13 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 /** @jsx JSXSlack.h */
 import { SectionBlock, MrkdwnElement } from '@slack/types'
 import { JSXSlack } from '../jsx'
-import { ObjectOutput } from '../utils'
+import { ObjectOutput, aliasTo } from '../utils'
 import html from '../html'
 import { BlockComponentProps } from './Blocks'
 import { Image } from './Image'
+import { Button } from './elements/Button'
+import { Select } from './elements/Select'
 
 export const Section: JSXSlack.FC<
   BlockComponentProps & { children: JSXSlack.Children<{}> }
@@ -51,6 +54,12 @@ export const Section: JSXSlack.FC<
         accessory = JSXSlack(
           <Image alt={child.props.alt} src={child.props.src} />
         )
+        eaten = true
+      } else if (child.type === 'button') {
+        accessory = JSXSlack(aliasTo(Button, child)!)
+        eaten = true
+      } else if (child.type === 'select') {
+        accessory = JSXSlack(aliasTo(Select, child)!)
         eaten = true
       }
     }