+++ src/block-kit/elements/Button.tsx
@@ -9,6 +9,7 @@ export interface ButtonProps {
   actionId?: string
   children: JSXSlack.Children<{}>
   confirm?: JSXSlack.Node<ConfirmProps>
+  name?: string
   style?: SlackButton['style']
   url?: string
   value?: string
@@ -18,7 +19,7 @@ export const Button: JSXSlack.FC<ButtonProps> = props => (
   <ObjectOutput<SlackButton>
     type="button"
     text={plainText(JSXSlack(<PlainText>{props.children}</PlainText>))}
-    action_id={props.actionId}
+    action_id={props.actionId || props.name}
     confirm={props.confirm ? JSXSlack(props.confirm) : undefined}
     style={props.style}
     url={props.url}